TINY-9957: Refactor Accordion key backspace/delete overrides
- move code from DetailsElement module to DetailsDelete module
- update DetailsDeleteTest to accommodate for refactor
